requirements.txt
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.6.0
psycopg[c]==3.2.6
django-environ==0.12.0
djangorestframework-simplejwt
django-extensions==3.2.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1